Updated slave garbage collection to do it based on disk usage (contributed by Vinod Kone, https://reviews.apache.org/r/6704).
git-svn-id: https://svn.apache.org/repos/asf/incubator/mesos/trunk@1386882 13f79535-47bb-0310-9956-ffa450edef68

+++ src/slave/gc.cpp
@@ -16,14 +16,18 @@
  * limitations under the License.
  */
 
+#include <map>
 #include <string>
+#include <vector>
 
 #include <process/delay.hpp>
 #include <process/dispatch.hpp>
 #include <process/future.hpp>
 #include <process/process.hpp>
+#include <process/timeout.hpp>
 
 #include <stout/duration.hpp>
+#include <stout/foreach.hpp>
 #include <stout/os.hpp>
 
 #include "logging/logging.hpp"
@@ -34,21 +38,56 @@ using namespace process;
 
 using process::wait; // Necessary on some OS's to disambiguate.
 
+using std::map;
 using std::string;
+using std::vector;
 
 namespace mesos {
 namespace internal {
 namespace slave {
 
+
 class GarbageCollectorProcess : public Process<GarbageCollectorProcess>
 {
 public:
+  virtual ~GarbageCollectorProcess();
+
   // GarbageCollector implementation.
   Future<bool> schedule(const Duration& d, const string& path);
 
+  void prune(const Duration& d);
+
 private:
-  void remove(const string& path, Promise<bool>* promise);
+  void remove(const Timeout& removalTime);
+
+  struct PathInfo
+  {
+    PathInfo(const string& _path, Promise<bool>* _promise)
+      : path(_path), promise(_promise) {}
+
+    string path;
+    Promise<bool>* promise;
+  };
+
+  // Store all the paths that needed to be deleted after a given timeout.
+  // NOTE: We are using std::map here instead of hashmap, because we
+  // need the keys of the map (deletion time) to be sorted in ascending order.
+  map<Timeout, vector<PathInfo> > paths;
+
+  void reset();
+  Timer timer;
 };
+
+
+GarbageCollectorProcess::~GarbageCollectorProcess()
+{
+  foreachvalue (const vector<PathInfo>& infos, paths) {
+    foreach (const PathInfo& info, infos) {
+      info.promise->future().discard();
+      delete info.promise;
+    }
+  }
+}
 
 
 Future<bool> GarbageCollectorProcess::schedule(
@@ -59,24 +98,77 @@ Future<bool> GarbageCollectorProcess::schedule(
 
   Promise<bool>* promise = new Promise<bool>();
 
-  delay(d, self(), &Self::remove, path, promise);
+  Timeout removalTime(d);
+
+  paths[removalTime].push_back(PathInfo(path, promise));
+
+  // If the timer is not yet initialized or the timeout is sooner than
+  // the currently active timer, update it.
+  if (timer.timeout().remaining() == Seconds(0) ||
+      removalTime < timer.timeout()) {
+    reset(); // Schedule the timer for next event.
+  }
 
   return promise->future();
 }
 
 
-void GarbageCollectorProcess::remove(
-    const string& path,
-    Promise<bool>* promise)
+// Fires a message to self for the next event. This also cancels any
+// existing timer.
+void GarbageCollectorProcess::reset()
 {
-  LOG(INFO) << "Removing " << path;
+  Timer::cancel(timer); // Cancel the existing timer, if any.
+  if (!paths.empty()) {
+    Timeout removalTime = (*paths.begin()).first; // Get the first entry.
+    Duration d = removalTime.remaining();
 
-  // TODO(benh): Check error conditions of 'rmdir', e.g., permission
-  // denied, file no longer exists, etc.
-  bool result = os::rmdir(path);
+    VLOG(1) << "Scheduling GC removal event to fire after " << d;
+    timer = delay(d, self(), &Self::remove, removalTime);
+  } else {
+    timer = Timer(); // Reset the timer.
+  }
+}
 
-  promise->set(result);
-  delete promise;
+
+void GarbageCollectorProcess::remove(const Timeout& removalTime)
+{
+  if (paths.count(removalTime) > 0) {
+    foreach (const PathInfo& info, paths[removalTime]) {
+      const string& path = info.path;
+      Promise<bool>* promise = info.promise;
+
+      LOG(INFO) << "Deleting " << path;
+
+      // TODO(benh): Check error conditions of 'rmdir', e.g., permission
+      // denied, file no longer exists, etc.
+      // TODO(vinod): Consider invoking rmdir via async.
+      bool result = os::rmdir(path);
+
+      VLOG(1) << "Deleted " << path;
+      promise->set(result);
+      delete promise;
+    }
+    paths.erase(removalTime);
+  } else {
+    // This might happen if the directory(s) has already been removed
+    // (e.g: by prune())
+    LOG(WARNING) << "Ignoring gc event at " << removalTime.remaining()
+                 << " as the corresponding directories are already removed";
+  }
+
+  reset(); // Schedule the timer for next event.
+}
+
+
+void GarbageCollectorProcess::prune(const Duration& d)
+{
+  foreachkey (const Timeout& removalTime, paths) {
+    if (removalTime.remaining() <= d) {
+      LOG(INFO) << "Pruning directories with remaining removal time "
+                << removalTime.remaining();
+      dispatch(self(), &GarbageCollectorProcess::remove, removalTime);
+    }
+  }
 }
 
 
@@ -91,6 +183,7 @@ GarbageCollector::~GarbageCollector()
 {
   terminate(process);
   wait(process);
+  delete process;
 }
 
 
@@ -99,6 +192,12 @@ Future<bool> GarbageCollector::schedule(
     const string& path)
 {
   return dispatch(process, &GarbageCollectorProcess::schedule, d, path);
+}
+
+
+void GarbageCollector::prune(const Duration& d)
+{
+  return dispatch(process, &GarbageCollectorProcess::prune, d);
 }
 
 } // namespace mesos {
